requirements.txt
# Wellness Agent Requirements for GitHub/Render Deployment

# FastAPI and web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
starlette>=0.40.0

# LangChain and AI/ML
langchain>=0.1.12
langchain-core>=0.1.32
langchain-community>=0.0.29
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.2
transformers>=4.36.0

# Vector database
faiss-cpu>=1.7.4

# Google Gemini API
google-generativeai>=0.3.2

# Data processing
pandas>=2.1.4
numpy>=1.24.3

# Time series forecasting (for wellness predictions)
prophet>=1.1.4
statsmodels>=0.14.0
scikit-learn>=1.3.0
scipy>=1.10.0

# PDF processing
pypdf>=3.17.4
PyPDF2>=3.0.1

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.2

# Logging and monitoring
structlog>=23.2.0

# Additional utilities
typing-extensions>=4.8.0
pyyaml>=6.0.1

# For production deployment
gunicorn>=21.2.0
